Cook class added
Added a cook class, there's still no reference to it from mainActivity, no way to open the camera and add a void cheque photo and no return value. Lets the user input the relevant data and stores it in variables.
Incomplete Cook class, still working on it but other people can reference this for further code for now.